fix(layered): compute Java's float arithmetic in double, as elkjs does
GWT compiles Java's float to a JavaScript number, so elkjs never rounds to
32 bits; elk-rs did, and its port positions after a node resize, its
orthogonal router's layer positions and its crossing minimization's port
barycenters and random perturbation drifted from elkjs by up to 1e-4.
These are now double, as in elkjs. On external/elk-models' 550 JSON models
this turns the three floating-point differences from elkjs into matches and
changes no other model's result.

Adds parity tests over every captured archboard fixture.

+++ plugins/org.eclipse.elk.alg.layered/src/org/eclipse/elk/alg/layered/graph/l_graph_util.rs
@@ -48,11 +48,11 @@ impl LGraphUtil {
             )
         };
 
-        // Java uses float (f32) for these variables. Truncate to match.
-        let width_ratio = (new_size.x / old_size.x) as f32 as f64;
-        let height_ratio = (new_size.y / old_size.y) as f32 as f64;
-        let width_diff = (new_size.x - old_size.x) as f32 as f64;
-        let height_diff = (new_size.y - old_size.y) as f32 as f64;
+        // Java declares these float; elkjs (GWT) computes them as doubles, and so does this.
+        let width_ratio = new_size.x / old_size.x;
+        let height_ratio = new_size.y / old_size.y;
+        let width_diff = new_size.x - old_size.x;
+        let height_diff = new_size.y - old_size.y;
 
         if move_ports {
             for port in ports {
